% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculate Greatest Common Divisor(GCD) using Euclidean algorithm.}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
GCD of \code{x} and \code{y}
}
\description{
Calculate Greatest Common Divisor(GCD) using Euclidean algorithm.
}
\details{
Purpose of this function is to find Greatest Common Divisor(GCD) among entered values using Euclidean algorithm.

In mathematics, the Euclidean algorithm is an efficient method for computing the greatest common divisor (GCD) of two numbers, the largest number that divides both of them without leaving a remainder.

Please refer below URL for more information on Algorithm 
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\examples{
\dontrun{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
}
